+++ hooks/utils.js
@@ -1,0 +1,11 @@
+"use strict"
+
+var fs = require("fs");
+
+function fileExists(filePath) {
+    return fs.existsSync(filePath);
+}
+
+module.exports = {
+    fileExists
+};